Do not discard top 8 significant bits of timestamp values
Instead of casting the 5-byte timestamps into 4-byte values, perform
as much of the arithmetic as 64-bit values instead. Also, since
apparently the GCC compiler for the ESP32 fails to preserve the sign
across division of int64_t values, remove the negative sign before
division and add it later. The sign preservation is required for
later timestamp debugging.